Resolves feedback: Refactors truncate function, rename number parser file, fixes changeset, improves notes template
