+++ packages/web-forms/src/lib/format/truncateDecimals.ts
@@ -1,18 +1,27 @@
-export const truncateDecimals = (num: number, decimals = 3): string => {
-	if (
-		typeof num !== 'number' ||
-		Number.isNaN(num) ||
-		typeof decimals !== 'number' ||
-		Number.isNaN(decimals)
-	) {
+interface TruncateDecimalsOptions {
+	readonly decimals: number;
+}
+
+export const truncateDecimals = (num: number, options: TruncateDecimalsOptions): string => {
+	const { decimals } = options;
+
+	if (!Number.isFinite(decimals) || decimals < 0) {
+		throw new Error(
+			`Truncation failed: expected a finite, non-negative decimal option, got: ${decimals}`
+		);
+	}
+
+	if (!Number.isFinite(num)) {
 		return '';
 	}
 
-	if (Number.isInteger(num) || decimals <= 0) {
+	if (Number.isInteger(num)) {
 		return num.toString();
 	}
 
 	const factor = Math.pow(10, decimals);
 	const withDecimals = Math.floor(Math.abs(num) * factor) / factor;
-	return (num < 0 ? -withDecimals : withDecimals).toString();
+	const sign = Math.sign(num);
+
+	return String(withDecimals * sign);
 };
